init save post feature

+++ InstagramClone/Models/FirebaseDB.cs
@@ -274,6 +274,15 @@ namespace InstagramClone.Models
         {
             try
             {
+                //Check for duplicate notication (spam case)
+                var result = (await FirebaseDB.firebaseClient
+                .Child("notification")
+                .Child(FirebaseDB.CurrentUserId)
+                .OnceAsync<NotificationModel>()).Where(item => item.Object.Type == noti.Type && item.Object.UserId == noti.UserId &&
+                (noti.Type == "follow" || (noti.Type == "postlike" && item.Object.PostId == noti.PostId)));
+
+                if (result != null) return;
+
                 await firebaseClient.Child("notification")
                     .Child(OwnerId)
                     .PostAsync(noti);
@@ -325,6 +334,8 @@ namespace InstagramClone.Models
                     }
                 });
         }
+
+
         public static async Task<string> InsertChatBox(string friendId)
         {
             UserChatboxModel userChatbox = new UserChatboxModel
@@ -410,6 +421,63 @@ namespace InstagramClone.Models
             await UpdateUserChatBox(receiverId, chatboxId, chat);
         }
 
+
+
+        public static async Task SavePost(PostModel post)
+        {
+            try
+            {
+                await firebaseClient
+                    .Child("savepost")
+                    .Child(CurrentUserId)
+                    .Child(post.PostId)
+                    .PutAsync(new PostModel
+                    {
+                        PostId = post.PostId,
+                        OwnerId = post.OwnerId,
+                        OwnerUsername = post.OwnerUsername,
+                        OwnerImage = post.OwnerImage,
+                        Caption = post.Caption,
+                        PostTime = post.PostTime
+                    });
+            }
+            catch(Exception e)
+            {
+                Console.WriteLine(e.Message);
+            }
+        }
+        public static async Task UnsavePost(PostModel post)
+        {
+            try
+            {
+                await firebaseClient
+                    .Child("savepost")
+                    .Child(CurrentUserId)
+                    .Child(post.PostId)
+                    .DeleteAsync();
+            }
+            catch(Exception e)
+            {
+                Console.WriteLine(e.Message);
+            }
+        }
+        public static async Task<bool> IsPostSaved(PostModel post)
+        {
+            try
+            {
+                var savePost = await firebaseClient
+                    .Child("savepost")
+                    .Child(CurrentUserId)
+                    .Child(post.PostId)
+                    .OnceSingleAsync<PostModel>();
+                return savePost != null;
+            }
+            catch(Exception e)
+            {
+                Console.WriteLine(e.Message);
+                return false;
+            }
+        }
         //End Nội
 
         //Dũng
@@ -525,16 +593,16 @@ namespace InstagramClone.Models
                     .Child("follower")
                     .Child(user2.UserKey)
                     .OnceAsync<FollowUser>()).Where(a => a.Object.UserKey == user1.UserKey).FirstOrDefault();
-
-                await firebaseClient.Child("following")
-                    .Child(user1.UserKey)
-                    .Child(toDeleteFollowing.Key)
-                    .DeleteAsync();
-
-                await firebaseClient.Child("follower")
-                    .Child(user2.UserKey)
-                    .Child(toDeleteFollower.Key)
-                    .DeleteAsync();
+                if (toDeleteFollowing != null)
+                    await firebaseClient.Child("following")
+                        .Child(user1.UserKey)
+                        .Child(toDeleteFollowing.Key)
+                        .DeleteAsync();
+                if (toDeleteFollower != null)
+                    await firebaseClient.Child("follower")
+                        .Child(user2.UserKey)
+                        .Child(toDeleteFollower.Key)
+                        .DeleteAsync();
 
                 return "unfollow";
             }
